+++ Changelog.md
@@ -2,8 +2,10 @@ RestrictEvents Changelog
 ========================
 #### v1.0.9
 - Added `revblock` for user configuration of blocking processes
-- Added optional `displaypolicyd` and `mediaanalysisd` blocking
-  - `revblock=gmux` and `revblock=media` options respectively
+- Added additional process blocking:
+  - `gmux` - block displaypolicyd on Big Sur+ (for genuine MacBookPro9,1/10,1)
+  - `media` - block mediaanalysisd on Ventura+ (for Metal 1 GPUs)
+  - `telemetry` - block telemetry plugin on Mojave+ (for SSE4,1 CPUs)
 
 #### v1.0.8
 - Added constants for macOS 13 support